4.1.8 Вынесет функции для отрисовки в отдельный файл
